+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Changed
 - Filters now hide on mobile and show in sidebar.  
 - Updated to Design System 0.1.0
+- Work Detail page has a Show All toggle
 
 ## [0.8.5]
 ### Added